unsigned_varint

Module encode

Basic unsigned-varint encoding.

Functions

u8

Encode the given u8 into the given byte array.

u8_buffer

Create new array buffer for encoding of u8 values.

u16

Encode the given u16 into the given byte array.

u32

Encode the given u32 into the given byte array.

u64

Encode the given u64 into the given byte array.

u16_buffer

Create new array buffer for encoding of u16 values.

u32_buffer

Create new array buffer for encoding of u32 values.

u64_buffer

Create new array buffer for encoding of u64 values.

u128

Encode the given u128 into the given byte array.

u128_buffer

Create new array buffer for encoding of u128 values.

usize

Encode the given usize into the given byte array.

usize_buffer

Create new array buffer for encoding of usize values.